chore(shop): more validation checks for employee and customer

+++ projects/shop/src/main/java/com/revature/model/Offer.java
@@ -40,7 +40,7 @@ public class Offer {
         return amount;
     }
 
-    public void setAmount(float amount) {
+    public void setAmount(double amount) {
         this.amount = amount;
     }
